api books and api author uploaded

+++ Day_11_First_Exercise/src/main/java/com/ariel/Controllers/UsersController.java
@@ -5,6 +5,7 @@ import com.ariel.Services.UsersService;
 
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
 import java.util.List;
 
 @Path("/users")
@@ -19,43 +20,59 @@ public class UsersController {
     @GET
     @Path("/get_users")
     @Produces(MediaType.APPLICATION_JSON)
-    public List<Users> getUsers(){
+    public Response getUsers(){
         List<Users> allUsers = usersService.getUsers();
-        return allUsers;
+        return Response.status(201).entity(allUsers).build();
     }
 
     @GET
     @Path("/get_user/{id}")
-    @Produces(MediaType.APPLICATION_JSON)
-    public Users getUser(@PathParam("id") int id){
+    @Produces("application/*")
+    public Response getUser(@PathParam("id") Integer id){
+        if(id == null){
+            String error = "Incomplete data";
+            return Response.status(400).entity(error).build();
+        }
         Users user = usersService.getUser(id);
-        return user;
+        return Response.status(201).entity(user).build();
     }
 
     @POST
     @Path("/save_user")
     @Consumes(MediaType.APPLICATION_JSON)
-    @Produces("application/text")
-    public String saveUser(Users user){
+    @Produces(MediaType.TEXT_PLAIN)
+    public Response saveUser(Users user){
+        if(user.getName() == null || user.getLastName() == null || user.getDni() == null){
+            String error = "Incomplete data";
+            return Response.status(400).entity(error).build();
+        }
         usersService.saveUser(user);
-        return "user saved";
+        return Response.status(Response.Status.ACCEPTED).entity("Saved person").build();
     }
 
     @PUT
     @Path("/update_user")
     @Consumes(MediaType.APPLICATION_JSON)
-    @Produces("application/text")
-    public String updateUser(Users user){
+    @Produces(MediaType.TEXT_PLAIN)
+    public Response updateUser(Users user){
+        if(user.getName() == null || user.getLastName() == null || user.getDni() == null || user.getId() == null){
+            String error = "Incomplete data";
+            return Response.status(400).entity(error).build();
+        }
         usersService.updateUser(user);
-        return "user updated";
+        return Response.status(201).entity("user updated").build();
     }
 
     @DELETE
     @Path("/delete_user/{id}")
-    @Produces("application/text")
-    public String deleteUser(@PathParam("id") int id){
+    @Produces(MediaType.TEXT_PLAIN)
+    public Response deleteUser(@PathParam("id") Integer id){
+        if(id == null){
+            String error = "Incomplete data";
+            return Response.status(400).entity(error).build();
+        }
         usersService.deleteUser(id);
-        return "user deleted";
+        return Response.status(201).entity("user deleted").build();
     }
 
 }
